Move code to ./src
